# openapi.yaml
openapi: 3.0.3
info:
  title: HotScan API
  description: |
    HotScan - 加密货币/DeFi 热点雷达 API
    
    实时监控加密资产交易信号、流动性变化和风险评估。
    
    ## 主要功能
    - 📊 交易信号查询和分析
    - 💎 加密资产详情和指标
    - 🤖 AI 驱动的术语问答（RAG）
    - 📱 分享海报生成
    - ⚙️ 后台任务触发（需要认证）
    
    ## 认证
    受保护的端点需要在请求头中提供 `X-Cron-Secret`。
    
  version: 2.0.0
  contact:
    name: HotScan API Support
    email: support@hotscan.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://hotscan.vercel.app
    description: Production Server
  - url: http://localhost:3000
    description: Development Server
...
paths:
  /api/signals:
    get:
      tags:
        - Signals
      summary: 获取交易信号列表
      description: |
        查询交易信号，按创建时间和风险分数降序排序。
        
        **排序规则**：
        1. createdAt DESC（最新优先）
        2. riskScore DESC（高风险优先）
        
        **性能优化**：使用复合索引 `(createdAt, riskScore)` 提升查询速度。
        
      operationId: getSignals
      parameters:
        - name: window
          in: query
          description: 时间窗口过滤
          required: false
          schema:
            type: string
            enum: [5m, 15m, 1h, 4h, 1d]
            example: 1h
        - name: limit
          in: query
          description: 返回结果数量
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: minRiskScore
          in: query
          description: 最低风险分数过滤
          required: false
          schema:
            type: number
            minimum: 0
            maximum: 100
            example: 60
        - name: assetSymbol
          in: query
          description: 资产符号过滤
          required: false
          schema:
            type: string
            example: BTC
        - name: sentiment
          in: query
          description: 市场情绪过滤
          required: false
          schema:
            type: string
            enum: [bullish, bearish, neutral]
            example: bullish
      responses:
        '200':
          description: 成功返回信号列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SignalWithAsset'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 125
                      limit:
                        type: integer
                        example: 20
                      window:
                        type: string
                        example: 1h
              examples:
                success:
                  summary: 成功响应示例
                  value:
                    success: true
                    data:
                      - id: clx123abc
                        assetId: clx456def
                        window: 1h
                        priceChangePct: 3.45
                        currentPrice: 67234.50
                        volZScore: 2.8
                        volumeUSD: 1250000000
                        liqDeltaPct: 1.2
                        totalLiquidityUSD: 850000000
                        top5HoldPct: 8.5
                        holderCount: 45678901
                        newWalletNetBuy: 340000
                        newWalletCount: 1234
                        riskScore: 35.5
                        contractAgeDays: 5475
                        sentiment: bullish
                        aiSummary: BTC稳健上涨，机构资金持续流入，风险较低
                        alertLevel: low
                        createdAt: '2024-10-01T12:00:00Z'
                        asset:
                          id: clx456def
                          symbol: BTC
                          name: Bitcoin
                          chain: bitcoin
                          logo: https://cryptologos.cc/logos/bitcoin-btc-logo.png
                          decimals: 8
                          isActive: true
                      - id: clx789ghi
                        assetId: clx012jkl
                        window: 1h
                        priceChangePct: 8.95
                        currentPrice: 0.000001234
                        volZScore: 4.2
                        volumeUSD: 5800000
                        liqDeltaPct: 3.8
                        totalLiquidityUSD: 28500000
                        top5HoldPct: 42.3
                        holderCount: 89456
                        newWalletNetBuy: 128000
                        newWalletCount: 567
                        riskScore: 68.5
                        contractAgeDays: 365
                        sentiment: bullish
                        aiSummary: PEPE出现暴涨信号，成交量激增，但需注意高风险和集中持仓
                        alertLevel: high
                        createdAt: '2024-10-01T11:45:00Z'
                        asset:
                          id: clx012jkl
                          symbol: PEPE
                          name: Pepe
                          chain: ethereum
                          logo: https://cryptologos.cc/logos/pepe-pepe-logo.png
                          decimals: 18
                          isActive: true
                    meta:
                      total: 125
                      limit: 20
                      window: 1h
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    CronSecret:
      type: apiKey
      in: header
      name: X-Cron-Secret
      description: |
        Cron 任务认证密钥。
        在环境变量 `CRON_SECRET` 中配置。
  schemas:
    Asset:
      type: object
      properties:
        id:
          type: string
          example: clx456def
        symbol:
          type: string
          example: BTC
        name:
          type: string
          example: Bitcoin
        chain:
          type: string
          example: bitcoin
        logo:
          type: string
          format: uri
          example: https://cryptologos.cc/logos/bitcoin-btc-logo.png
        decimals:
          type: integer
          example: 8
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Signal:
      type: object
      properties:
        id:
          type: string
        assetId:
          type: string
        window:
          type: string
          enum: [5m, 15m, 1h, 4h, 1d]
        priceChangePct:
          type: number
          description: 价格变化百分比
        currentPrice:
          type: number
          description: 当前价格
        volZScore:
          type: number
          description: 成交量 Z-Score
        volumeUSD:
          type: number
          description: 成交量（美元）
        liqDeltaPct:
          type: number
          description: 流动性变化百分比
        totalLiquidityUSD:
          type: number
          description: 总流动性（美元）
        top5HoldPct:
          type: number
          description: 前5大持仓占比
        holderCount:
          type: integer
          description: 持仓地址数
        newWalletNetBuy:
          type: number
          description: 新钱包净买入（美元）
        newWalletCount:
          type: integer
          description: 新钱包数量
        riskScore:
          type: number
          minimum: 0
          maximum: 100
          description: 风险评分
        contractAgeDays:
          type: integer
          description: 合约年龄（天）
        sentiment:
          type: string
          enum: [bullish, bearish, neutral]
          description: 市场情绪
        aiSummary:
          type: string
          description: AI 分析摘要
        alertLevel:
          type: string
          enum: [low, medium, high, critical]
          description: 警报级别
        createdAt:
          type: string
          format: date-time
    SignalWithAsset:
      allOf:
        - $ref: '#/components/schemas/Signal'
        - type: object
          properties:
            asset:
              $ref: '#/components/schemas/Asset'
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: 错误代码
            message:
              type: string
              description: 错误消息
            details:
              type: object
              description: 详细信息
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
              description: 错误时间
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation:
              summary: 参数验证失败
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: 请求参数验证失败
                  details:
                    field: window
                    constraint: 必须是 5m, 15m, 1h, 4h, 1d 之一
                  timestamp: '2024-10-01T12:00:00Z'
            missing:
              summary: 缺少必需参数
              value:
                success: false
                error:
                  code: MISSING_PARAMETER
                  message: 缺少必需参数
                  details:
                    field: q
                    type: string
                  timestamp: '2024-10-01T12:00:00Z'
...
    TooManyRequests:
      description: 请求频率超限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            rateLimit:
              summary: 超过频率限制
              value:
                success: false
                error:
                  code: RATE_LIMIT_EXCEEDED
                  message: 请求频率超过限制
                  details:
                    limit: 100
                    window: 60
                    retryAfter: 30
                  timestamp: '2024-10-01T12:00:00Z'
    InternalServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            server:
              summary: 服务器错误
              value:
                success: false
                error:
                  code: INTERNAL_SERVER_ERROR
                  message: 服务器内部错误，请稍后重试
                  details:
                    requestId: req_abc123
                  timestamp: '2024-10-01T12:00:00Z'
            database:
              summary: 数据库错误
              value:
                success: false
                error:
                  code: DATABASE_ERROR
                  message: 数据库连接失败
                  timestamp: '2024-10-01T12:00:00Z'